Extract URL parsing logic to UrlParser
